+++ backend/associations/fixtures/page.yaml
@@ -1,13 +1,16 @@
 - model: associations.page
   pk: 1
   fields:
-    title: La tarte aux pommes
-    text: Lorem ipsum dolor sit amet, consectetur adipisicing elit. Accusantium atque
-      eos fugiat fugit laudantium magnam molestiae, odit omnis, quam reprehenderit,
-      repudiandae suscipit ullam voluptatem? Consequuntur eius eveniet optio perferendis
-      quisquam! Alias laborum maiores nemo optio perspiciatis porro quaerat sapiente
-      sequi ut voluptatum? Cupiditate deserunt error hic illum impedit iusto molestiae
-      placeat quasi quidem quis quod recusandae repellat sint, veniam voluptates.
+    title: Showdown
+    text:
+      "Showdown is a Javascript Markdown to HTML converter, based on the original works by John Gruber. It can be used client side (in the browser) or server side (with Node or io).
+      # Installation
+      ## Download tarball
+      You can download the latest release tarball directly from [releases][releases]
+      ## Bower
+      bower install showdown
+      ## npm (server-side)
+      npm install showdown"
     association: bde
-    creation_date: '2018-08-18T13:19:37+00:00'
-    last_update_date: '2018-08-18T13:19:37+00:00'
+    creation_date: "2018-08-18T13:19:37+00:00"
+    last_update_date: "2018-08-18T13:19:37+00:00"
