+++ frontend/package.json
@@ -16,6 +16,7 @@
         "react-debounce-input": "^3.2.2",
         "react-dom": "^16.13.1",
         "react-dropzone": "^11.0.1",
+        "react-markdown": "^4.3.1",
         "react-query": "1.2.1",
         "react-router-dom": "^5.1.2",
         "react-scripts": "^3.4.1",
